+++ mockredis/__init__.py
@@ -1,3 +1,3 @@
-from mockredis.redis import MockRedis, mock_redis_client, mock_strict_redis_client
+from mockredis.client import MockRedis, mock_redis_client, mock_strict_redis_client
 
 __all__ = ["MockRedis", "mock_redis_client", "mock_strict_redis_client"]